requirements.txt
# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.7
arabic-reshaper==3.0.0
python-bidi==0.4.2

# Translation
googletrans==4.0.0rc1

# Utilities
tqdm==4.66.1
pillow==10.1.0
